+++ src/components/Auth/RegisterSuccess.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom";
-import Candy from "../Header/Navigator/Candy" ;
+import Candy from "../layout/Header/Navigator/Candy";
 import { monsterDefault } from "@/assets/monster";
 
 type RegisterSuccessProps = {
@@ -8,7 +8,10 @@ type RegisterSuccessProps = {
   avatar: string;
 };
 
-export default function RegisterSuccess({ nickname, avatar }: RegisterSuccessProps) {
+export default function RegisterSuccess({
+  nickname,
+  avatar,
+}: RegisterSuccessProps) {
   const navigate = useNavigate();
   const [seconds, setSeconds] = useState(5);
 
@@ -29,16 +32,22 @@ export default function RegisterSuccess({ nickname, avatar }: RegisterSuccessPro
 
   return (
     <div className="w-full h-screen md:w-auto  md:h-auto dark shadow-2xl rounded-lg p-6 bg-[var(--card)]  flex flex-col items-center justify-center gap-4">
-        <div className="flex items-end -gap-10 justify-center">
-          <img src={avatar} alt="選擇的頭像" className="w-60 rounded-full mx-auto my-4" />
-          <img src={monsterDefault} alt="monster" className="w-32 h-36 -ml-24" />
-        </div>
+      <div className="flex items-end -gap-10 justify-center">
+        <img
+          src={avatar}
+          alt="選擇的頭像"
+          className="w-60 rounded-full mx-auto my-4"
+        />
+        <img src={monsterDefault} alt="monster" className="w-32 h-36 -ml-24" />
+      </div>
       <div className=" flex flex-col items-center justify-center gap-2">
         <h2 className="text-lg font-bold ">歡迎 {nickname} 進入樂園！</h2>
-        <p className="text-sm border-gradient-set-1">快拿糖果去參加試煉吧，減肥為了吃更多糖果</p>
+        <p className="text-sm border-gradient-set-1">
+          快拿糖果去參加試煉吧，減肥為了吃更多糖果
+        </p>
       </div>
-        <Candy amount={10000} />
-        <p className="text-sm text-gray-500">{seconds}s後自動跳轉首頁</p>
+      <Candy amount={10000} />
+      <p className="text-sm text-gray-500">{seconds}s後自動跳轉首頁</p>
     </div>
   );
 }
